added functionality as per pt 3 of specs.
still need to sort the sell and buying price deal.

Idea that just popped up: items that a store buys isnt filled with
items, more over an HashMap of an item name to the price that they will
pay for them, kinda makes everything a bit more simple.

or even better. HashMap mapping an item name to its qualities (price,
and spaceTaken). then an item is constructed based on these qualities
and then sold to player. doesnt really make sense to do it otherwise, bc
there is essentially going to be the same item in both store possession
that they can sell, and in ships possession. which is basically
superpositon, and i know this project isnt on quantum physics.


HashMap of HashMap?, explained key item is an item name, and then its
values is another hash map. this way it makes it more clear when
indexing and getting items for displaying, if you get what i mean.
otherwise it may turn into some arbitary indexing convention for things
in the value part of hashmap.

if this is implemented, then the whole display method needs to be
changed..

However, i may be overly obsessive over such a small detail, lmk