Added more visitStore functionality to the Store Class
added static methods as below

added a getDisplayArrayList(catalogue) method that adds list
functionality to the option thing that we were talking about.

also another method for finding the name of a chosen item.

gameEnvironment is more becoming like a collection of objects, most of
the functionality i can roll back to the core object classes.

we need a way to handle exceptions, should this be in GE or in ui?