Updates to Store
Further developed the transaction system.

added methods buyItemChecker and sellItemChecker(), which help with the
cmd line for finding the reasons for why a sale didnt go right.

previously in the past, i was limited to passing around booleans. so i
decided to handle this in the most top level of selling and buying
things before GE and cmd line.

let me know if its bad style to return null or a string. used to be
boolean, but i see no other way to integrate, or if there is, it would
require alot of work, that i just dont think is worth it right now.

to note is that some of the code that exists in cmd line is only
temporary in location, i intend to move some of this into GE, or even
into GameUi (if we make this abstract), which would make more sense bc
GE shoudl be treated like a collection of objects and not a place to get
methods.
this will make code alot easier to maintain across versions, which is
always a plus, easier to read, and more efficient! may be more effort
that its worth. and perhaps we are going overboard.

let me know what you think.



may be some bad style. however, it works, and i dont see any other work
around this!