Changes to ship class
upgrades are now added to their own array, stored sperately from Items.
this makes alot of sense with pirate attacks, who can only take Items,
but not instances of their child class, ShipUpgrade.

fixed a bug in takeDamage(), where division was casting to an int, and
not staying at a float, revealed in unit testing.