Changes to Ship class
added a new parameter in constructor called ship size. this is the value
of the crewsize, and the max cargo space is now size *10

also added a thing called max DefenseCapability which makes alot more
sense, and less effort than having a remainign upgrade space for a ship

upgrades now no longer count against quota of ship items, the number of
upgrades that you have is limitedby the max defense Capability

also added a method for seeing the upgrades that have been bought for a
ship. this is used by ui when asked.

in description of ship, now includes maxDefenseStat too.