Changes to visitStoreScreen
initialized alot of methods with buying and selling.

i dont think this is going to be all that hard to implement. we just
need to implement it. all the logic is in core classes, and we can use
cmd line ui as a starting block for any other logic that may be needed