JavaDocing and code editing
Added a bunch of Javadoc for class attributes.

Edited some method javadocs to make sure that they all look the same
style throughout the project. When I saw javadoc that was inaccurate i
changed them too, ie had descriptions for inputs that were no longer
there. we need to do this properly and go over the entire  project.

Also need to make sure that when you do javadoc, you NEED to make sure
that you include the type of the variable. Explicitly say this directly
after the input variable name.this is in the specs.

dont need to include variable tuype for class attributes.

Reformatted some classes to just make them a bit more neat an readable.
Ie putting getters and setters to the bottom. And grouping similar
methods together.