Updates to Screen
Making it as functional as possible. It now sets all the characteristics
of Screen that will be common to all screens, but leaves setting bounds
to you.

Will probably make an abstract class initialize() which the conssrtuctor
will call, would be a bit nicer and every screen will need that anyway.