Changes to GE
added method setMinMoneyToTravel() which is useful for testing.

may be unneccessary to have this. Check later. Assume that it works
properly. and in store tests i can just make up values. Check with
these.

And if i need to, check that it is getting the right values from game
environment.