Changes to Store *kinda big*
store now doesnt check if selling an item puts them into liquidation
before buying, and then denies sale if so. Now just does this for
upgrades only.

As a convention, price of items that a store buys is the same that they
sell. This makes logic a bit more easier to do. Variations in price is
left to other stores to sort out. So that still needs to be done in
balancing.

denies buying an upgrade if it puts a user into liquidation, because
there is no way to sell that upgrade back!