Changes to store test
added test cases to test the new code in Store that i just added. May be
a good idea to look at these test cases, to make sure that they are
doing a good job of testing the code, however, this is extra, and i feel
like there are enough test cases already.

coverage for store is around 60% this is because there are alot of
getters and setters, plus alot of methods based around catalogues that
dont really need to be changed (wont exactly break running of program...
yet).